Work with exception rules
- Work with multiple lists
- Update tests
- Improve override graphs dir handling
- Standardize dir names (sugarcoated_scripts -> sugarcoat_scripts)
- Add examples

+++ sugarcoat-pipeline.js
@@ -14,7 +14,7 @@ const execFileAsync = promisify(execFile);
 const defaultCrawlSecs = 30;
 const defaultPolicyJson = 'policy.json';
 const defaultOutputDir = 'output';
-const maxRetries = 5;
+const defaultSamples = 3;
 
 // Parser options
 const parser = new argparseLib.ArgumentParser({
@@ -37,9 +37,10 @@ parser.add_argument('-d', '--debug', {
   action: 'store_true',
   default: false,
 });
-parser.add_argument('-l', '--filter-list', {
-  help: 'Filter list to use',
+parser.add_argument('-l', '--filter-lists', {
+  help: 'Filter lists to use',
   required: true,
+  nargs: '+',
 });
 parser.add_argument('-p', '--policy', {
   help: `Path to policy file. Default: ${defaultPolicyJson}`,
@@ -58,57 +59,100 @@ parser.add_argument('-k', '--keep', {
   action: 'store_true',
   default: false,
 });
+parser.add_argument('-s', '--samples', {
+  help: `Number of times a URL is attempted to be crawled. Default: ${defaultSamples}`,
+  type: 'int',
+  default: defaultSamples,
+});
 
 const args = parser.parse_args();
 const binary = args.binary;
 const crawlUrl = args.url;
 const policyJsonFile = args.policy;
-const filterlist = args.filter_list;
+const filterlists = args.filter_lists;
 const debug = args.debug;
 const secs = args.secs;
 const keep = args.keep;
+const samples = args.samples;
 const graphsDirOverride = args.graphs_dir_override ? path.resolve(args.graphs_dir_override) : null;
+// Directory paths
 const outputDir = path.resolve(args.output);
 const graphsDir = path.join(outputDir, '/graphs');
 const scriptsDir = path.join(outputDir, '/scripts');
-const sugarcoatedScriptsDir = path.join(outputDir, '/sugarcoated_scripts');
-const massagedConfigJson = path.join(outputDir, '/config.json');
+const sugarcoatedScriptsDir = path.join(outputDir, '/sugarcoat_scripts');
+// Generated SugarCoat files paths
+const massagedConfigJsonFile = path.join(outputDir, '/config.json');
+const traceFile = path.join(outputDir, '/sugarcoat_trace.json');
+const reportFile = path.join(outputDir, '/sugarcoat_report.html');
+const rulesFile = path.join(outputDir, '/sugarcoat_rules.txt');
+const resourcesFile = path.join(outputDir, '/sugarcoat_resources.json');
+
 let scriptNameToUrl = {}; // Used to get 'patterns' URL for each script
+let urlsSeen = new Set(); // We don't want to store duplicate scripts
 
-// Always clean up at start
-const preCleanup = async () => {
-  // Check if policy JSON file exists
-  await fs.access(policyJsonFile, constants.F_OK).catch(() => {
-    const errMsg = policyJsonFile + ' not found!';
-    throw new Error(errMsg);
-  });
-  debug && console.debug('Cleaning up generated directories...');
-  // Remove generated directory if it exists and create new
-  await fs.rmdir(outputDir, { force: true, recursive: true });
-  await fs.mkdir(outputDir, { recursive: true });
-  fs.mkdir(scriptsDir, { recursive: true });
-  if (!graphsDirOverride) fs.mkdir(graphsDir, { recursive: true });
+/**
+ * Helper functions
+ */
+const checkFileExistence = async file => {
+  return fs.access(file, constants.F_OK);
 };
-
 const readGraphFiles = async graphsDir => {
   const files = await fs.readdir(graphsDir);
   return files.filter(file => path.extname(file).toLowerCase() === '.graphml');
 };
 
+/**
+ * Pipeline functions
+ */
+
+// Always clean up at start
+const preCheckAndClean = async () => {
+  // Check if policy + filterlist files exist
+  await Promise.all(
+    [...filterlists, policyJsonFile].map(async file => checkFileExistence(file))
+  ).catch(err => {
+    const errMsg = err.path + ' not found!';
+    throw new Error(errMsg);
+  });
+  debug && console.debug('Cleaning up generated directories...');
+  // clean sugarcoated_scripts, scripts, config.json, report, resources, rules, trace
+  await Promise.all([
+    ...[traceFile, reportFile, resourcesFile, massagedConfigJsonFile, rulesFile].map(
+      async file => {
+        fs.unlink(file).catch(_ => {
+          return;
+        });
+      }
+    ),
+    ...[sugarcoatedScriptsDir, scriptsDir].map(async dir =>
+      fs.rmdir(dir, { force: true, recursive: true })
+    ),
+  ]);
+  // and if not graphs dir override then graphs too
+  if (graphsDirOverride !== graphsDir) {
+    await fs.rmdir(graphsDir, { force: true, recursive: true });
+  }
+  // Check if outputDir exists. If yes, then leave it be. If no, create it.
+  await checkFileExistence(outputDir).catch(_ => fs.mkdir(outputDir, { recursive: true }));
+  // Create scripts and graphs dir
+  fs.mkdir(scriptsDir, { recursive: true });
+  if (!graphsDirOverride) fs.mkdir(graphsDir, { recursive: true });
+};
+
 const generateGraphs = async (retriesLeft, graphsDir, readLocal) => {
   let errorMsg;
-  if (readLocal) {
+  if (readLocal || retriesLeft == 0) {
+    // Graph files should exist at this point. If not, then error.
     const graphFiles = await readGraphFiles(graphsDir);
     if (graphFiles.length == 0) {
       errorMsg = 'No files found in ' + graphsDir + ' that end with .graphml';
       throw new Error(errorMsg);
     }
+    debug && console.debug(graphFiles);
+    debug && console.debug('Pagegraph-crawl done!');
     return graphFiles;
   }
-  if (retriesLeft == 0) {
-    errorMsg = 'Tried generating graphs using pagegraph-crawl ' + maxRetries + 'x but failed!';
-    throw new Error(errorMsg);
-  }
+
   if (!binary || !crawlUrl) {
     errorMsg =
       'Must provide path to PageGraph-enabled browser binary (via --binary)' +
@@ -120,22 +164,15 @@ const generateGraphs = async (retriesLeft, graphsDir, readLocal) => {
     binary +
     '" --secs ' +
     secs +
+    ' --shields down' +
     ' --url ' +
     crawlUrl +
     ' --output ' +
     graphsDir;
   debug && console.debug('Running pagegraph-crawl with command: ' + cmd);
   await execAsync(cmd);
-  const graphFiles = await readGraphFiles(graphsDir);
-  debug && console.log(graphFiles);
-  if (graphFiles.length == 0) {
-    const newRetriesLeft = retriesLeft - 1;
-    debug && console.debug('No files found in ' + graphsDir + ' that end with .graphml');
-    debug && console.debug('Retries left: ' + newRetriesLeft);
-    return generateGraphs(newRetriesLeft, graphsDir, readLocal);
-  }
-  debug && console.debug('Pagegraph-crawl finished running!');
-  return graphFiles;
+  const newRetriesLeft = retriesLeft - 1;
+  return generateGraphs(newRetriesLeft, graphsDir, readLocal);
 };
 
 const getSources = async (graphFiles, graphsDir) => {
@@ -148,56 +185,80 @@ const getSources = async (graphFiles, graphsDir) => {
       );
       const pagegraphBinaryArgs = ['-f', path.join(graphsDir, graphFile)];
       const options = { windowsHide: true };
-      const adblockArgs = [...pagegraphBinaryArgs, 'adblock_rules', '-l', filterlist];
-
-      let cmdOutput;
-      try {
-        cmdOutput = await execFileAsync(pagegraphBinary, adblockArgs, options);
-      } catch (err) {
-        return; // if there is a weirdly-shaped domain, don't error out
-      }
-
-      let jsonOutput = JSON.parse(cmdOutput.stdout);
-      const edges = jsonOutput.flatMap(edge =>
-        edge.requests.map(requestAndEdgeTuple => requestAndEdgeTuple[1])
-      );
-      // For each edge that corresponds to script, get downstream requests
-      const requests = (
-        await Promise.all(
-          edges.flatMap(async edge => {
-            const downstreamRequestsArgs = [
-              ...pagegraphBinaryArgs,
-              'downstream_requests',
-              edge,
-              '--requests',
-            ];
-            cmdOutput = await execFileAsync(pagegraphBinary, downstreamRequestsArgs, options);
-            jsonOutput = JSON.parse(cmdOutput.stdout);
-            return jsonOutput;
-          })
-        )
-      ).flat();
-      const uniqueRequests = [...new Set(requests)];
-      // For each request id, get the source and put into outputDir
       await Promise.all(
-        uniqueRequests.map(async requestId => {
-          const requestIdInfoArgs = [...pagegraphBinaryArgs, 'request_id_info', requestId];
+        filterlists.map(async filterlist => {
+          const adblockArgs = [...pagegraphBinaryArgs, 'adblock_rules', '-l', filterlist];
+          let cmdOutput;
           try {
-            cmdOutput = await execFileAsync(pagegraphBinary, requestIdInfoArgs, options);
+            cmdOutput = await execFileAsync(pagegraphBinary, adblockArgs, options);
           } catch (err) {
-            return; // if request ID is not related to script, the rust binary returns error code
+            return; // if there is a weirdly-shaped domain, don't error out
           }
-          jsonOutput = JSON.parse(cmdOutput.stdout);
-          const origUrl = jsonOutput.url;
-          const parsedUrl = new URL(origUrl);
-          let scriptName = path.basename(parsedUrl.pathname, '.js');
-          let source = jsonOutput.source;
-          const scriptFilePath = path.join(scriptsDir, scriptName + '.js');
-          let unusedScriptFilename = await unusedFilename(scriptFilePath, {
-            incrementer: unusedFilename.separatorIncrementer('-'),
-          });
-          scriptNameToUrl[path.basename(unusedScriptFilename, '.js')] = origUrl;
-          fs.writeFile(unusedScriptFilename, source, { recursive: true });
+          let jsonOutput = JSON.parse(cmdOutput.stdout);
+          const edges = jsonOutput
+            .flatMap(edge =>
+              edge.requests.map(request => {
+                if (request.blocking_filter && request.exception_filter) {
+                  return request.edge_id;
+                }
+              })
+            )
+            .filter(x => x); // to filter out nulls
+          // For each edge that corresponds to script, get downstream requests
+          const requests = (
+            await Promise.all(
+              edges.flatMap(async edge => {
+                const downstreamRequestsArgs = [
+                  ...pagegraphBinaryArgs,
+                  'downstream_requests',
+                  edge,
+                  '--requests',
+                ];
+                cmdOutput = await execFileAsync(pagegraphBinary, downstreamRequestsArgs, options);
+                jsonOutput = JSON.parse(cmdOutput.stdout);
+                debug &&
+                  console.log(
+                    'requests ' +
+                      jsonOutput +
+                      ' for edge ' +
+                      edge +
+                      ' for graph ' +
+                      graphFile +
+                      ' for list ' +
+                      filterlist
+                  );
+                return jsonOutput;
+              })
+            )
+          ).flat();
+
+          const uniqueRequests = [...new Set(requests)];
+          // For each request id, get the source and put into outputDir
+          await Promise.all(
+            uniqueRequests.map(async requestId => {
+              const requestIdInfoArgs = [...pagegraphBinaryArgs, 'request_id_info', requestId];
+              try {
+                cmdOutput = await execFileAsync(pagegraphBinary, requestIdInfoArgs, options);
+              } catch (err) {
+                return; // if request ID is not related to script, the rust binary returns error code
+              }
+              jsonOutput = JSON.parse(cmdOutput.stdout);
+              const origUrl = jsonOutput.url;
+              if (urlsSeen.has(origUrl)) {
+                return;
+              }
+              const parsedUrl = new URL(origUrl);
+              let scriptName = path.basename(parsedUrl.pathname, '.js');
+              let source = jsonOutput.source;
+              const scriptFilePath = path.join(scriptsDir, scriptName + '.js');
+              let unusedScriptFilename = await unusedFilename(scriptFilePath, {
+                incrementer: unusedFilename.separatorIncrementer('_'),
+              });
+              scriptNameToUrl[path.basename(unusedScriptFilename, '.js')] = origUrl;
+              fs.writeFile(unusedScriptFilename, source, { recursive: true });
+              urlsSeen.add(origUrl);
+            })
+          );
         })
       );
     })
@@ -211,11 +272,11 @@ const massageConfig = async graphs => {
   const policy = config.policy;
   config.graphs = [path.join(graphs, '/*.graphml')];
   config.code = scriptsDir;
-  config.trace = path.join(outputDir, '/sugarcoat_trace.json');
-  config.report = path.join(outputDir, '/sugarcoat_report.html');
+  config.trace = traceFile;
+  config.report = reportFile;
   const bundle = {
-    rules: path.join(outputDir, '/sugarcoat_rules.txt'),
-    resources: path.join(outputDir, '/sugarcoat_resources.json'),
+    rules: rulesFile,
+    resources: resourcesFile,
   };
   config.bundle = bundle;
   config.targets = {};
@@ -229,8 +290,7 @@ const massageConfig = async graphs => {
     config.targets[targetKey] = newObj;
   });
   debug && console.debug('Writing massaged config.json... ');
-  debug && console.dir(config, { depth: null });
-  await fs.writeFile(massagedConfigJson, JSON.stringify(config), { recursive: true });
+  await fs.writeFile(massagedConfigJsonFile, JSON.stringify(config), { recursive: true });
 };
 
 const runSugarCoat = async () => {
@@ -238,7 +298,7 @@ const runSugarCoat = async () => {
     'node ' +
     path.join('node_modules', 'sugarcoat', 'cli.js') +
     ' --config ' +
-    massagedConfigJson +
+    massagedConfigJsonFile +
     ' --ingest --report --rewrite --bundle';
   debug && console.debug('Running sugarcoat with command: ' + cmd);
   await execAsync(cmd);
@@ -270,16 +330,21 @@ const postCleanup = async () => {
   }
 };
 
+/**
+ * Run pipeline
+ */
 (async () => {
-  await preCleanup();
+  await preCheckAndClean();
   const graphsDirToUse = graphsDirOverride ? graphsDirOverride : graphsDir;
   const readLocal = !!graphsDirOverride;
-  const graphFiles = await generateGraphs(maxRetries, graphsDirToUse, readLocal);
+  const graphFiles = await generateGraphs(samples, graphsDirToUse, readLocal);
   await getSources(graphFiles, graphsDirToUse);
   await massageConfig(graphsDirToUse);
   await runSugarCoat();
   await postCleanup();
 })().catch(err => {
   console.error(err.message);
-  fs.rmdir(outputDir, { force: true, recursive: true });
+  if (!keep) {
+    fs.rmdir(outputDir, { force: true, recursive: true });
+  }
 });
